move utility to jssdk
